requirements.txt
langchain
gradio
streamlit
openai
tiktoken
python-dotenv
pinecone-client
langchain-community
langchain-experimental
langchain-openai
pandas
langchain-core
watchdog
tabulate